Main window operations
Add commands to set a main window, and focus it, as well as swap the
current window with the main window. Main window is per-workspace, and
is only special in terms of the move and focus targets. If in the main
window, and a different window had been swapped into the main spot, the
main will switch with the last window that was in the main spot.